+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to
 
 ### Added
 
+- Projects page welcome section: allow users to learn how to use the app thru
+  Arcade videos [#2563](https://github.com/OpenFn/lightning/issues/2563)
+- Store user preferences in database
+  [#2564](https://github.com/OpenFn/lightning/issues/2564)
+
 ### Changed
 
 ### Fixed